@@ -25,86 +27,74 @@
     "precommit": "lint-staged"
   },
   "dependencies": {
-    "prop-types": "^15.5.10",
-    "react-autowhatever": "^10.1.0",
-    "shallow-equal": "^1.0.0"
+    "es6-promise": "^4.2.8",
+    "prop-types": "^15.7.2",
+    "react-themeable": "^1.1.0",
+    "section-iterator": "^2.0.0",
+    "shallow-equal": "^1.2.1"
   },
   "peerDependencies": {
-    "react": ">=0.14.7"
+    "react": ">=16.3.0"
   },
   "devDependencies": {
-    "autoprefixer": "^6.7.7",
-    "autosuggest-highlight": "^3.1.0",
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.25.0",
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "^6.4.1",
-    "babel-plugin-react-transform": "^2.0.2",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.6",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-register": "^6.24.1",
-    "chai": "^3.5.0",
-    "cross-env": "^3.1.4",
-    "css-loader": "^0.28.4",
-    "es6-promise": "^4.1.0",
-    "eslint": "^3.19.0",
-    "eslint-plugin-react": "6.10.3",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.11.2",
-    "husky": "^0.13.4",
-    "ismobilejs": "^0.4.1",
+    "@babel/cli": "^7.8.4",
+    "@babel/core": "^7.9.0",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/preset-env": "^7.9.0",
+    "@babel/preset-react": "^7.9.4",
+    "@babel/register": "^7.9.0",
+    "@istanbuljs/nyc-config-babel": "^3.0.0",
+    "autoprefixer": "^9.7.5",
+    "autosuggest-highlight": "^3.1.1",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.1.0",
+    "babel-plugin-istanbul": "^6.0.0",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
+    "chai": "^4.2.0",
+    "cross-env": "^7.0.2",
+    "css-loader": "^3.4.2",
+    "eslint": "^6.8.0",
+    "eslint-plugin-react": "^7.19.0",
+    "file-loader": "^6.0.0",
+    "husky": "^4.2.3",
+    "ismobilejs": "^1.0.3",
     "isomorphic-fetch": "^2.2.1",
-    "jsdom": "^9.12.0",
-    "less": "^2.7.2",
-    "less-loader": "^2.2.3",
-    "lint-staged": "^3.6.1",
-    "mocha": "^3.4.2",
-    "nyc": "^10.1.2",
+    "jsdom": "15.1.1",
+    "less": "^3.11.1",
+    "less-loader": "^5.0.0",
+    "lint-staged": "^10.0.10",
+    "mini-css-extract-plugin": "^0.9.0",
+    "mocha": "^7.1.1",
+    "nyc": "^15.0.0",
     "openurl": "^1.1.1",
-    "postcss-loader": "^1.3.3",
-    "prettier": "1.7.4",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
-    "react-modal": "^1.7.7",
+    "postcss-loader": "^3.0.0",
+    "prettier": "^2.0.2",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-modal": "^3.11.2",
     "react-transform-hmr": "^1.0.4",
-    "sinon": "^1.17.7",
-    "sinon-chai": "^2.11.0",
-    "style-loader": "^0.18.2",
-    "svgo": "^0.7.2",
-    "svgo-loader": "^1.2.1",
-    "url-loader": "^0.5.9",
-    "webpack": "^1.14.0",
-    "webpack-dev-server": "^1.16.2"
+    "sinon": "^9.0.1",
+    "sinon-chai": "^3.5.0",
+    "style-loader": "^1.1.3",
+    "svgo": "^1.3.2",
+    "svgo-loader": "^2.2.1",
+    "terser-webpack-plugin": "^2.3.5",
+    "url-loader": "^4.0.0",
+    "webpack": "^4.42.1",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3"
   },
   "files": [
     "dist"
   ],
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
   "lint-staged": {
-    ".*.js": [
-      "npm run prettier",
-      "git add"
-    ],
     "*.js": [
-      "npm run prettier",
-      "git add"
-    ],
-    "demo/src/**/*.js": [
-      "npm run prettier",
-      "git add"
-    ],
-    "demo/standalone/app.js": [
-      "npm run prettier",
-      "git add"
-    ],
-    "src/**/*.js": [
-      "npm run prettier",
-      "git add"
-    ],
-    "test/**/*.js": [
-      "npm run prettier",
-      "git add"
+      "npm run prettier:src"
     ]
   },
   "keywords": [
