@@ -4,44 +4,54 @@ import Autosuggest from '../../src/Autosuggest';
 import languages from '../plain-list/languages';
 import { escapeRegexCharacters } from '../../demo/src/components/utils/utils.js';
 
-const getMatchingLanguages = value => {
+const getMatchingLanguages = (value) => {
   const escapedValue = escapeRegexCharacters(value.trim());
   const regex = new RegExp('^' + escapedValue, 'i');
 
-  return languages.filter(language => regex.test(language.name));
+  return languages.filter((language) => regex.test(language.name));
 };
 
 let app = null;
 
-export const getSuggestionValue = sinon.spy(suggestion => {
+export const setHighlightFirstSuggestion = (value) => {
+  app.setState({
+    highlightFirstSuggestion: value,
+  });
+};
+
+export const getSuggestionValue = sinon.spy((suggestion) => {
   return suggestion.name;
 });
 
-export const renderSuggestion = sinon.spy(suggestion => {
+export const renderSuggestion = sinon.spy((suggestion) => {
   return <span>{suggestion.name}</span>;
 });
 
 export const onChange = sinon.spy((event, { newValue }) => {
   app.setState({
-    value: newValue
+    value: newValue,
   });
 });
 
 export const onSuggestionsFetchRequested = sinon.spy(({ value }) => {
   app.setState({
-    suggestions: getMatchingLanguages(value)
+    suggestions: getMatchingLanguages(value),
   });
 });
 
 export const onSuggestionsClearRequested = sinon.spy(() => {
   app.setState({
-    suggestions: []
+    suggestions: [],
   });
 });
 
 export const onSuggestionSelected = sinon.spy();
 
-export const onSuggestionHighlighted = sinon.spy();
+export const onSuggestionHighlighted = sinon.spy(({ suggestion }) => {
+  app.setState({
+    highlightedSuggestion: suggestion,
+  });
+});
 
 export default class AutosuggestApp extends Component {
   constructor() {
@@ -51,15 +61,17 @@ export default class AutosuggestApp extends Component {
 
     this.state = {
       value: '',
-      suggestions: []
+      suggestions: [],
+      highlightedSuggestion: null,
+      highlightFirstSuggestion: false,
     };
   }
 
   render() {
-    const { value, suggestions } = this.state;
+    const { value, suggestions, highlightFirstSuggestion } = this.state;
     const inputProps = {
       value,
-      onChange
+      onChange,
     };
 
     return (
@@ -72,7 +84,7 @@ export default class AutosuggestApp extends Component {
         getSuggestionValue={getSuggestionValue}
         renderSuggestion={renderSuggestion}
         inputProps={inputProps}
-        highlightFirstSuggestion={true}
+        highlightFirstSuggestion={highlightFirstSuggestion}
       />
     );
   }
