@@ -1,5 +1,7 @@
-import jsdom from 'jsdom';
+import { JSDOM } from 'jsdom';
 
-global.document = jsdom.jsdom('<!doctype html><html><body></body></html>');
-global.window = global.document.defaultView;
-global.navigator = global.window.navigator;
+const dom = new JSDOM(`<!doctype html><html><body></body></html>`);
+
+global.window = dom.window;
+global.document = dom.window.document;
+global.navigator = dom.window.navigator;
